+++ DelvUI/Interface/Nameplates/NameplateConfig.cs
@@ -515,6 +515,14 @@ namespace DelvUI.Interface.GeneralElements
         [Order(3)]
         public bool DisableInteraction = false;
 
+        [Checkbox("Use Different Size when targeted", spacing = true)]
+        [Order(31)]
+        public bool UseDifferentSizeWhenTargeted = false;
+
+        [DragInt2("Size When Targeted", min = 1, max = 4000)]
+        [Order(32, collapseWith = nameof(UseDifferentSizeWhenTargeted))]
+        public Vector2 SizeWhenTargeted;
+
         [ColorEdit4("Targeted Border Color")]
         [Order(38, collapseWith = nameof(DrawBorder))]
         public PluginConfigColor TargetedBorderColor = PluginConfigColor.FromHex(0xFFFFFFFF);
@@ -561,6 +569,11 @@ namespace DelvUI.Interface.GeneralElements
         public bool IsVisible(uint hp, uint maxHp)
         {
             return Enabled && (!OnlyShowWhenNotFull || hp < maxHp) && !(HideHealthAtZero && hp <= 0);
+        }
+
+        public Vector2 GetSize(bool targeted)
+        {
+            return targeted && UseDifferentSizeWhenTargeted ? SizeWhenTargeted : Size;
         }
 
         public NameplateBarConfig() : base(Vector2.Zero, Vector2.Zero, new(Vector4.Zero)) { } // don't remove
@@ -722,6 +735,11 @@ namespace DelvUI.Interface.GeneralElements
             {
                 bar.FillColor = PluginConfigColor.FromHex(bgColor);
                 bar.BackgroundColor = PluginConfigColor.FromHex(0xAA000000);
+            }
+
+            if (barConfig is NameplateBarConfig nameplateBar)
+            {
+                nameplateBar.SizeWhenTargeted = nameplateBar.Size;
             }
 
             EditableLabelConfig nameLabelConfig = new EditableLabelConfig(new Vector2(0, -20), "[name]", DrawAnchor.Top, DrawAnchor.Bottom)